chore(Canvas): Split `CanvasService` in `ControllerService` and `FlowService`
Split `CanvasService` for better separation of concerns.

fix: https://github.com/KaotoIO/kaoto/issues/1329
